+++ src/screens/TaskScreen/Task/TaskCategory.tsx
@@ -1,3 +1,5 @@
+import { useState } from 'react'
+import { Pressable } from 'react-native'
 import { View } from 'tamagui'
 
 import BottomSheet, { Option } from '@/components/BottomSheet'
@@ -8,8 +10,10 @@ import { customPalettes } from '@/theme/customPalettes'
 import useToggle from '@/util/useToggle'
 import ThreeDots from '@assets/Svgs/ThreeDots.svg'
 
+const options = ['Sprint A', 'Sprint B', 'Sprint C']
 const TaskCategory = () => {
   const [open, setOpen] = useToggle(false)
+  const [category, setCategory] = useState<string>('Sprint A')
   return (
     <View w="100%">
       <Task.Context onPress={setOpen}>
@@ -17,17 +21,19 @@ const TaskCategory = () => {
           <Typography type="R" fontSize={14} textColor={customPalettes.gray[700]}>
             카테고리
           </Typography>
-          <Tag color="dark">Sprint A</Tag>
+          <Tag color="dark">{category}</Tag>
         </View>
         <ThreeDots />
       </Task.Context>
       <BottomSheet isOpen={open} bottomSheetTitle="카테고리" setOpen={setOpen}>
-        {Array.from({ length: 3 }).map((_, index) => (
-          <Option key={index}>
+        {options.map(item => (
+          <Option key={item}>
             <Option.Text type="R" fontSize={16} textColor={customPalettes.gray[900]}>
-              스프린트 A
+              {item}
             </Option.Text>
-            <Option.Radio checked />
+            <Pressable onPress={() => setCategory(item)}>
+              <Option.Radio checked={category === item} />
+            </Pressable>
           </Option>
         ))}
       </BottomSheet>
